// src/obniz/libs/io_peripherals/common.d.ts
export type DriveType = "5v" | "3v" | "open-drain";
export type PullType = "5v" | "3v" | "0v";
export type AnimationStatus = "loop" | "registrate" | "pause" | "resume";
export type StopBitType = 1 | 1.5 | 2;
export type BitType = 5 | 6 | 7 | 8;
export type ParityType = "off" | "odd" | "even";
export type FlowControlType = "off" | "rts" | "cts" | "rts-cts";